proste wyjatki,

+++ src/main/java/com/example/demo/service/LessonService.java
@@ -1,8 +1,6 @@
 package com.example.demo.service;
 
-import com.example.demo.model.Lesson;
-import com.example.demo.model.Student;
-import com.example.demo.model.Teacher;
+import com.example.demo.model.*;
 import com.example.demo.repository.LessonRepository;
 import com.example.demo.repository.StudentRepository;
 import com.example.demo.repository.TeacherRepository;
@@ -25,20 +23,23 @@ public class LessonService {
     }
 
     public void deleteById(long id) {
-        lessonRepository.deleteById(id);
+        Lesson lesson = lessonRepository.findById(id).orElseThrow(() -> new LessonNotFoundException("Lesson not found: " + id));
+        if (lesson.getDatetime().isAfter(LocalDateTime.now())) {
+            lessonRepository.deleteById(id);
+        }
     }
 
     public void save(Lesson lesson, Long teacherId, Long studentId) {
         Teacher teacher = teacherRepository.findById(teacherId)
-                .orElseThrow(() -> new RuntimeException("Teacher not found: " + teacherId));
+                .orElseThrow(() -> new TeacherNotFoundException("Teacher not found: " + teacherId));
         Student student = studentRepository.findById(studentId)
-                .orElseThrow(() -> new RuntimeException("Student not found: " + studentId));
+                .orElseThrow(() -> new StudentNotFoundException("Student not found: " + studentId));
 
         LocalDateTime from = lesson.getDatetime().minusHours(1);
         LocalDateTime to = lesson.getDatetime().plusHours(1);
 
-        if(lessonRepository.existsByTeacherAndDatetimeGreaterThanAndDatetimeLessThan(teacher, from, to)) {
-            throw new RuntimeException("Lekcja juz istnieje!");
+        if (lessonRepository.existsByTeacherAndDatetimeGreaterThanAndDatetimeLessThan(teacher, from, to)) {
+            throw new LessonAlreadyExists("Lekcja juz istnieje!");
         }
         lesson.setTeacher(teacher);
         lesson.setStudent(student);
